feat: implement organization deletion functionality and refactor related components
