+++ sdk/longlink/identity.py
@@ -1,6 +1,7 @@
 import jwt
 from uuid import UUID
-from datetime import UTC, datetime, timedelta
+from datetime import timedelta
+from longlink.utils.time import utcnow
 
 IDENTITY_TOKEN_ALGORITHM = "HS256"
 IDENTITY_TOKEN_AUDIENCE = "longlink:identity"
@@ -11,7 +12,7 @@ def create_identity_token(user_id: UUID, secret: str) -> str:
     """Create one short-lived Platform identity assertion for a Solution request."""
 
     # Bind the assertion to one user and a fixed runtime-only token purpose.
-    issued_at = datetime.now(UTC)
+    issued_at = utcnow()
     return jwt.encode(
         {
             "sub": str(user_id),
